workspace/requirements.txt
pyspark==3.2.4
google-cloud-bigquery
google-cloud-pubsub